Read Urls from text
